Add "add_html_id_to_subsection" function for manual html_ids
Previously, our html_id attributes would be added when we called "save",
but now we're saving time by not calling "save".

That's great, but it causes problems because it means the "html_id"
is not getting added to the subsection.

So now I have refactored the function so that we call it within
the create_nofo function rather than relying on the "save" function
to do it for is.

This means we can save time and still keep our heading links.